+++ example/transpose/transpose_test.c
@@ -16,8 +16,8 @@ int main(int argc, char const *argv[])
         17, 18, 19, 20,
         21, 22, 23, 24
     };
-    int shapeA[] = {2, 3, 4};
-    int dimA = 3;
+    long shapeA[] = {2, 3, 4};
+    long dimA = 3;
 
     // Print A
     onnx_tensor_info(A, shapeA, dimA);
@@ -27,8 +27,8 @@ int main(int argc, char const *argv[])
     float* B = onnx_tensor_transpose(A, shapeA, dimA, perm);
 
     // Print B
-    int shapeB[] = {4, 2, 3};
-    int dimB = 3;
+    long shapeB[] = {4, 2, 3};
+    long dimB = 3;
     onnx_tensor_info(B, shapeB, dimB);
 
     // Free memory
